pyproject.toml
[project]
name = "komuzik"
version = "0.1.0"
description = ""
authors = [{ name = "Kotaz", email = "semechkagent@gmail.com" }]
requires-python = ">=3.10,<4.0"
readme = "README.md"
dependencies = [
    "yt-dlp (>=2025.3.31,<2026.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pyrogram[asyncio] (>=2.0.0,<3.0.0)",
    "pyyaml (>=6.0,<7.0)",
    "tgcrypto",
    "telethon (>=1.33.0,<2.0.0)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
